Worked on HW3 some more with video capture in OpenCV with C++.
I got the video capture to work and was able to convert the image and output it in real time. I tried doing the SVD in real time but that just did not work well. The image displayed but was horribly laggy. Could be this POS laptop. Will try out on the desktop when I get time and a cheap webcam for it.