Rearranged the thesis folder structure.
I ran crack detection tests on both the nylon and steel rods. I wrote a matlab file to generate figures for the tests. The matlab file takes 5 cracked and 5 uncracked data sets for a particular rod. It then finds the difference from the first uncracked to successive samples and plots it. You can clearly see where the defect is introduced. It also plots an averaging of the 5 data sets for the cracked and the 5 data sets for the uncracked. This should give some extra material for my thesis and I think all that is left is to just put it all together. Bet that I can have it all done by next week. F yea!!!