Been playing around with parallel programming. Apparently OpenCV no longer supports OpenMP officially and has instead switched to Intels TBB or Threaded Building Blocks. I found an example program online and have been fiddling around with it. I dont seem to have fatals.h or hrtimer.h installed on my system which seems to be something that should be ther by default. Oh well, not a bit deal. Other ways to time.
I was timing using ctime library and the clock() function. I found, however, that when using multi-cores the clock functions returns the number of clocks multiplied by the number of cores used so need to watch out for this. Using time.h library and gettimeofday function to do the timing has confirmed this.

I have also found that version 2.4 of OpenCV has a SUBSTANTIALLY slower SVD function than 2.2 version of OpenCV that can be installed via the package manager in Fedora 17