Worked more on the Computer Vision Program 1.
Think I have the basic feature detection algorithm implemented. Just need to play around with it some more and make sure that it is working correctly.

Decided to not use the opencv Trace and Determinant functions since we just have a 2x2 matrix to do it on which is trivial