Got the suppressNonMaximumAdaptive function to work with dstVec instead of dstMat. Was making stupid errors but it forced me to find other errors. Just putting statements in the wrong for loop and what not. Oh well.
Seems to be working good. Added in the draw circles feature to get a better idea of the features.

The suppressNonMaximumAdaptive now returns a vectore containing the features in sorted ascending order. Now, you pass in a numKeep value to the suppression function and it will keep the top numKeep values based on intensity.

The feature vector contains a struct called featVal which will contain various feature properties such as location, scale, major orientation, minor orientation (for feature matching), intensity response value, and other properties.